don't rm the symlink.

+++ autogen.sh
@@ -7,15 +7,16 @@ test -z "$srcdir" && srcdir=.
 PKG_NAME="gnome-mud"
 
 (test -f $srcdir/configure.ac) || {
-    echo -n "**Error**: Directory "\`$srcdir\'" does not look like the"
+    echo -n "**Error**: Directory "\'$srcdir\'" does not look like the"
     echo " top-level $PKG_NAME directory"
     exit 1
 }
 
-if [ -f configure.in ]; then
+if [ -f configure.in -a ! -L configure.in ]; then
     echo "configure.in is not a symlink to configure.ac."
     echo "Move it out of the way and try again."
-elif [ ! -L configure.in ]; then
+    exit 1
+else
     rm -f configure.in
     ln -s configure.ac configure.in
 fi
@@ -26,5 +27,3 @@ which gnome-autogen.sh || {
 }
 
 USE_GNOME2_MACROS=1 . gnome-autogen.sh
-
-rm -f configure.in
